requirements.txt
allure-pytest==2.8.10
coverage==5.0.3
numpy==1.18.1
openpyxl==3.0.3
pandas==1.0.1
pillow==7.0.0
PyAutoGUI==0.9.48
PyAutoIt==0.6.0
pyjavaproperties==0.7
pylint==2.5.0
pytest==5.3.5
pytest-xdist==1.31.0
selenium==3.141.0
Sphinx==3.0.3
xlrd==1.2.0
